pyproject.toml
[project]
name = "posting"
version = "1.0.0b33"
description = "The modern API client that lives in your terminal."
authors = [
    { name = "Darren Burns", email = "darrenb900@gmail.com" }
]
dependencies = [
    "click>=8.1.7",
    "xdg-base-dirs>=6.0.1",
    "click-default-group>=1.2.4",
    "httpx[brotli]>=0.27.0",
    "textual-autocomplete>=3.0.0a3",
    "pyperclip>=1.8.2",
    "pydantic>=2.7.3",
    "pyyaml>=6.0.1",
    "textual[syntax]==0.70.0",
    "pydantic-settings>=2.3.4",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.11"
license = { file = "LICENSE" }
keywords = ["tui", "http", "client", "terminal", "api", "testing", "textual", "cli", "posting", "developer-tool"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/darrenburns/posting"
Repository = "https://github.com/darrenburns/posting"
Issues = "https://github.com/darrenburns/posting/issues"
Documentation = "https://github.com/darrenburns/posting/blob/main/README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
posting = "posting.__main__:cli"

[tool.rye]
managed = true
dev-dependencies = [
    "textual-dev>=1.5.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/posting"]
